% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modules.R
\name{create_modules}
\alias{create_modules}
\title{Create Shiny module files}
\usage{
create_modules()
}
\value{
This function does not return any value.
}
\description{
This function allows users to create multiple Shiny module files quickly and easily. 
Users can specify the path where they want to save the files, the number 
of files they want to create, and the content type for each file. The 
function supports two content types: 'default' and 'named'. 
For the 'default' content type, the module template uses default object names. 
For the 'named' content type, the module template uses object names which 
includes the file name.
}
\details{
This function prompts the user to provide:
\itemize{
  \item The path where they want to save the files.
  \item The number of files they want to create.
  \item The names of the files.
  \item The content type for each file ('default' or 'named').
}
}
